bindings: manifests uses entities/types
