+++ pkg/domain/entities/manifest.go
@@ -1,6 +1,9 @@
 package entities
 
-import "github.com/containers/image/v5/types"
+import (
+	"github.com/containers/image/v5/types"
+	entitiesTypes "github.com/containers/podman/v4/pkg/domain/entities/types"
+)
 
 // ManifestCreateOptions provides model for creating manifest
 type ManifestCreateOptions struct {
@@ -79,14 +82,7 @@ type ManifestModifyOptions struct {
 // ManifestPushReport provides the model for the pushed manifest
 //
 // swagger:model
-type ManifestPushReport struct {
-	// ID of the pushed manifest
-	ID string `json:"Id"`
-	// Stream used to provide push progress
-	Stream string `json:"stream,omitempty"`
-	// Error contains text of errors from pushing
-	Error string `json:"error,omitempty"`
-}
+type ManifestPushReport = entitiesTypes.ManifestPushReport
 
 // ManifestRemoveOptions provides the model for removing digests from a manifest
 //
@@ -97,26 +93,9 @@ type ManifestRemoveOptions struct {
 // ManifestRemoveReport provides the model for the removed manifest
 //
 // swagger:model
-type ManifestRemoveReport struct {
-	// Deleted manifest list.
-	Deleted []string `json:",omitempty"`
-	// Untagged images. Can be longer than Deleted.
-	Untagged []string `json:",omitempty"`
-	// Errors associated with operation
-	Errors []string `json:",omitempty"`
-	// ExitCode describes the exit codes as described in the `podman rmi`
-	// man page.
-	ExitCode int
-}
+type ManifestRemoveReport = entitiesTypes.ManifestRemoveReport
 
 // ManifestModifyReport provides the model for removed digests and changed manifest
 //
 // swagger:model
-type ManifestModifyReport struct {
-	// Manifest List ID
-	ID string `json:"Id"`
-	// Images to removed from manifest list, otherwise not provided.
-	Images []string `json:"images,omitempty" schema:"images"`
-	// Errors associated with operation
-	Errors []error `json:"errors,omitempty"`
-}
+type ManifestModifyReport = entitiesTypes.ManifestModifyReport
